+++ example/advection/advection_driver.cpp
@@ -20,8 +20,10 @@
 #include "advection_package.hpp"
 #include "bvals/cc/bvals_cc_in_one.hpp"
 #include "interface/metadata.hpp"
+#include "interface/update.hpp"
 #include "mesh/meshblock_pack.hpp"
 #include "parthenon/driver.hpp"
+#include "refinement/refinement.hpp"
 
 using namespace parthenon::driver::prelude;
 
@@ -55,10 +57,12 @@ TaskStatus UpdateMeshData(const int stage, Integrator *integrator,
                           std::shared_ptr<parthenon::MeshData<Real>> &base,
                           std::shared_ptr<parthenon::MeshData<Real>> &dudt,
                           std::shared_ptr<parthenon::MeshData<Real>> &out) {
+  Kokkos::Profiling::pushRegion("Task_Advection_UpdateMeshData");
   const Real beta = integrator->beta[stage - 1];
   const Real dt = integrator->dt;
-  parthenon::Update::AverageMeshData(in, base, beta);
-  parthenon::Update::UpdateMeshData(in, dudt, beta * dt, out);
+  parthenon::Update::AverageIndependentData(in, base, beta);
+  parthenon::Update::UpdateIndependentData(in, dudt, beta * dt, out);
+  Kokkos::Profiling::popRegion(); // Task_Advection_UpdateMeshData
   return TaskStatus::complete;
 }
 
@@ -118,7 +122,8 @@ TaskCollection AdvectionDriver::MakeTaskCollection(BlockList_t &blocks, const in
     auto &mdudt = pmesh->mesh_data.GetOrAdd("dUdt", i);
 
     // compute the divergence of fluxes of conserved variables
-    auto flux_div = tl.AddTask(none, parthenon::Update::FluxDivergenceMesh, mc0, mdudt);
+    auto flux_div =
+        tl.AddTask(none, parthenon::Update::FluxDivergence<MeshData<Real>>, mc0, mdudt);
 
     // apply du/dt to all independent fields in the container
     auto update =
@@ -174,6 +179,7 @@ TaskCollection AdvectionDriver::MakeTaskCollection(BlockList_t &blocks, const in
   }
 
   TaskRegion &async_region2 = tc.AddRegion(num_task_lists_executed_independently);
+
   for (int i = 0; i < blocks.size(); i++) {
     auto &pmb = blocks[i];
     auto &tl = async_region2[i];
@@ -182,41 +188,28 @@ TaskCollection AdvectionDriver::MakeTaskCollection(BlockList_t &blocks, const in
     auto clear_comm_flags = tl.AddTask(none, &MeshBlockData<Real>::ClearBoundary,
                                        sc1.get(), BoundaryCommSubset::all);
 
-    auto prolongBound = tl.AddTask(
-        none,
-        [](std::shared_ptr<MeshBlock> pmb) {
-          pmb->pbval->ProlongateBoundaries(0.0, 0.0);
-          return TaskStatus::complete;
-        },
-        pmb);
+    auto prolongBound = none;
+    if (pmesh->multilevel) {
+      prolongBound = tl.AddTask(none, parthenon::ProlongateBoundaries, sc1);
+    }
 
     // set physical boundaries
     auto set_bc = tl.AddTask(prolongBound, parthenon::ApplyBoundaryConditions, sc1);
 
     // fill in derived fields
     auto fill_derived =
-        tl.AddTask(set_bc, parthenon::FillDerivedVariables::FillDerived, sc1);
+        tl.AddTask(set_bc, parthenon::Update::FillDerived<MeshBlockData<Real>>, sc1);
 
     // estimate next time step
     if (stage == integrator->nstages) {
       auto new_dt = tl.AddTask(
-          fill_derived,
-          [](std::shared_ptr<MeshBlockData<Real>> &rc) {
-            auto pmb = rc->GetBlockPointer();
-            pmb->SetBlockTimestep(parthenon::Update::EstimateTimestep(rc));
-            return TaskStatus::complete;
-          },
-          sc1);
+          fill_derived, parthenon::Update::EstimateTimestep<MeshBlockData<Real>>, sc1);
 
       // Update refinement
       if (pmesh->adaptive) {
+        using tag_type = TaskStatus(std::shared_ptr<MeshBlockData<Real>> &);
         auto tag_refine = tl.AddTask(
-            fill_derived,
-            [](std::shared_ptr<MeshBlock> pmb) {
-              pmb->pmr->CheckRefinementCondition();
-              return TaskStatus::complete;
-            },
-            pmb);
+            fill_derived, static_cast<tag_type *>(parthenon::Refinement::Tag), sc1);
       }
     }
   }