+++ src/bvals/bvals_interfaces.hpp
@@ -82,21 +82,11 @@ enum {
 };
 #endif
 
-// identifiers for all 6 faces of a MeshBlock
-enum BoundaryFace {
-  undef = -1,
-  inner_x1 = 0,
-  outer_x1 = 1,
-  inner_x2 = 2,
-  outer_x2 = 3,
-  inner_x3 = 4,
-  outer_x3 = 5
-};
 // TODO(felker): BoundaryFace must be unscoped enum, for now. Its enumerators are used as
 // int to index raw arrays (not ParArrayNDs)--> enumerator vals are explicitly specified
 
 // identifiers for boundary conditions
-enum class BoundaryFlag { block = -1, undef, reflect, outflow, periodic };
+enum class BoundaryFlag { block = -1, undef, reflect, outflow, periodic, user };
 
 // identifiers for types of neighbor blocks (connectivity with current MeshBlock)
 enum class NeighborConnect {